Reorganized files in unit testing project
