+++ 1.Domain/Domain.UnitTests/Dummy/Builders/Dummy/DummyContractBuilder.cs
@@ -1,49 +1,45 @@
 ﻿using Domain.Services.Contracts.Seed;
 using System;
 
-namespace Domain.Services.Impl.IntegrationTests.Dummy.Builders.Dummy
+namespace Domain.Services.Impl.UnitTests.Dummy.Builders.Dummy
 {
-    using Dummy = Model.Seed.Dummy;
-
-    internal class DummyBuilder : IBuilder<Dummy>
+    internal class DummyContractBuilder : IBuilder<CreateDummyContract>
     {
-        private Guid id;
         private string name;
         private string description;
         private string testValue;
 
-        public DummyBuilder()
+        public DummyContractBuilder()
         {
-            id = Guid.NewGuid();
+            var id = Guid.NewGuid();
             name = $"Test {id}";
             description = $"this is a dymmy for {name}";
             testValue = "Test value";
         }
 
-        public Dummy Build()
+        public CreateDummyContract Build()
         {
-            return new Dummy()
+            return new CreateDummyContract()
             {
-                Id = id,
                 Name = name,
                 Description = description,
                 TestValue = testValue
             };
         }
 
-        internal DummyBuilder WithName(string name)
+        internal DummyContractBuilder WithName(string name)
         {
             this.name = name;
             return this;
         }
 
-        internal DummyBuilder WithDescription(string description)
+        internal DummyContractBuilder WithDescription(string description)
         {
             this.description = description;
             return this;
         }
 
-        internal DummyBuilder WithTestValue(string testValue)
+        internal DummyContractBuilder WithTestValue(string testValue)
         {
             this.testValue = testValue;
             return this;
